refactor: 3 layer 기반에서 DDD 기반의 프로젝트 구조 변경
