+++ src/main/java/re/kr/icuh/icuhplatformadmin/article/api/request/ApproveEditRequest.java
@@ -1,0 +1,4 @@
+package re.kr.icuh.icuhplatformadmin.article.api.request;
+
+public class ApproveEditRequest {
+}
