+++ src/main/java/re/kr/icuh/icuhplatformadmin/article/api/response/ArticleListResponse.java
@@ -1,0 +1,28 @@
+package re.kr.icuh.icuhplatformadmin.article.api.response;
+
+import re.kr.icuh.icuhplatformadmin.article.domain.Article;
+import re.kr.icuh.icuhplatformadmin.article.domain.ArticleStatus;
+
+import java.time.LocalDateTime;
+
+public record ArticleListResponse (
+        Long id,
+        String title,
+        String authorOrganization,
+        String department,
+        String author,
+        ArticleStatus status,
+        LocalDateTime createdAt
+) {
+    public static ArticleListResponse fromEntity(Article article) {
+        return new ArticleListResponse(
+                article.getId(),
+                article.getTitle(),
+                article.getAuthorOrganization(),
+                article.getDepartment(),
+                article.getAuthor(),
+                article.getStatus(),
+                article.getCreatedAt()
+        );
+    }
+}
