+++ src/main/java/re/kr/icuh/icuhplatformadmin/article/domain/ArticleStatus.java
@@ -1,0 +1,11 @@
+package re.kr.icuh.icuhplatformadmin.article.domain;
+
+public enum ArticleStatus {
+    PENDING,
+    APPROVED,
+    REJECTED,
+    DELETED,
+    UPDATED_PENDING,
+    UPDATED_APPROVED,
+    DELETED_PENDING
+}
