+++ src/main/java/re/kr/icuh/icuhplatformadmin/article/domain/DocumentType.java
@@ -1,0 +1,25 @@
+package re.kr.icuh.icuhplatformadmin.article.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "document_types")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class DocumentType {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "en_name")
+    private String enName;
+}
