--- src/main/java/re/kr/icuh/icuhplatformadmin/config/QuerydslConfig.java
@@ -1,19 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class QuerydslConfig {
-
-    private final EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory queryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
