--- src/main/java/re/kr/icuh/icuhplatformadmin/controller/ArticleController.java
@@ -1,42 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
-@Controller
-public class ArticleController {
-
-	@GetMapping("/article/create-list")
-	public String createList() {
-		return "article/article-create-list";
-	}
-
-	@GetMapping("/article/update-list")
-	public String updateList() {
-		return "article/article-update-list";
-	}
-
-	@GetMapping("/article/delete-list")
-	public String deleteList() {
-		return "article/article-delete-list";
-	}
-
-    // 생성 요청 상세보기
-	@GetMapping("/article/detail/{id}")
-	public String articleDetail(@PathVariable("id") Long id) {
-		return "article/article-create-detail";
-	}
-
-    // 삭제 요청 상세보기
-    @GetMapping("/article/delete-detail/{id}")
-    public String articleDeleteDetail(@PathVariable("id") Long id) {
-        return "article/article-delete-detail";
-    }
-
-    // 수정 요청 상세보기
-    @GetMapping("/article/update-detail/{id}")
-    public String articleUpdateDetail(@PathVariable("id") Long id) {
-        return "article/article-update-detail";
-    }
-}
