--- src/main/java/re/kr/icuh/icuhplatformadmin/domain/article/ArticleEditRequest.java
@@ -1,92 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.domain.article;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.testcontainers.shaded.com.google.common.hash.Hashing;
-import re.kr.icuh.icuhplatformadmin.domain.file.FileEditRequest;
-
-import java.nio.charset.StandardCharsets;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "article_edit_requests")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class ArticleEditRequest {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "article_id")
-    private Article article;
-
-    @Column(name = "title")
-    private String title;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "author")
-    private String author;
-
-    @Column(name = "author_organization")
-    private String authorOrganization;
-
-    @Column(name = "department")
-    private String department;
-
-    @Column(name = "temp_password")
-    private String tempPassword;
-
-    @CreationTimestamp
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "views")
-    private Integer views;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status")
-    private ArticleStatus status;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "document_type_id")
-    private DocumentType documentType;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "subject_domain_id")
-    private SubjectDomain subjectDomain;
-
-    @Column(name = "source")
-    private String source;
-
-    @OneToMany(mappedBy = "articleEditRequest", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<FileEditRequest> files = new ArrayList<>();
-
-    @Builder
-    public ArticleEditRequest(String title, Article article, String description, String author, String authorOrganization, String department, String tempPassword, Integer views, ArticleStatus status, DocumentType documentType, SubjectDomain subjectDomain, String source) {
-        this.title = title;
-        this.article = article;
-        this.description = description;
-        this.author = author;
-        this.authorOrganization = authorOrganization;
-        this.department = department;
-        this.tempPassword = sha256Encode(tempPassword);
-        this.views = views == null ? 0 : views;
-        this.status = status == null ? ArticleStatus.UPDATED_PENDING : status;
-        this.documentType = documentType;
-        this.subjectDomain = subjectDomain;
-        this.source = source;
-    }
-
-    public String sha256Encode(String tempPassword) {
-        return Hashing.sha256().hashString(tempPassword, StandardCharsets.UTF_8).toString();
-    }
-
-}
