--- src/main/java/re/kr/icuh/icuhplatformadmin/domain/file/FileStatus.java
@@ -1,10 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.domain.file;
-
-public enum FileStatus {
-    PENDING,
-    APPROVED,
-    REJECTED,
-    DELETED,
-    UPDATED_PENDING,
-    DELETED_PENDING
-}
