--- src/main/java/re/kr/icuh/icuhplatformadmin/dto/ArticleListResponse.java
@@ -1,28 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.dto;
-
-import re.kr.icuh.icuhplatformadmin.domain.article.Article;
-import re.kr.icuh.icuhplatformadmin.domain.article.ArticleStatus;
-
-import java.time.LocalDateTime;
-
-public record ArticleListResponse (
-        Long id,
-        String title,
-        String authorOrganization,
-        String department,
-        String author,
-        ArticleStatus status,
-        LocalDateTime createdAt
-) {
-    public static ArticleListResponse fromEntity(Article article) {
-        return new ArticleListResponse(
-                article.getId(),
-                article.getTitle(),
-                article.getAuthorOrganization(),
-                article.getDepartment(),
-                article.getAuthor(),
-                article.getStatus(),
-                article.getCreatedAt()
-        );
-    }
-}
