--- src/main/java/re/kr/icuh/icuhplatformadmin/dto/ArticlePendingResponse.java
@@ -1,44 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.dto;
-
-import re.kr.icuh.icuhplatformadmin.domain.article.Article;
-import re.kr.icuh.icuhplatformadmin.domain.article.ArticleStatus;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public record ArticlePendingResponse(
-        Long id,
-        String title,
-        String description,
-        String author,
-        String authorOrganization,
-        String department,
-        ArticleStatus status,
-        LocalDateTime createdAt,
-        LocalDateTime updatedAt,
-        Integer views,
-        DocumentTypeResponse classification,
-        SubjectDomainResponse serviceType,
-        List<FileResponse> files
-) {
-    public static ArticlePendingResponse fromEntity(Article article) {
-        return new ArticlePendingResponse(
-                article.getId(),
-                article.getTitle(),
-                article.getDescription(),
-                article.getAuthor(),
-                article.getAuthorOrganization(),
-                article.getDepartment(),
-                article.getStatus(),
-                article.getCreatedAt(),
-                article.getUpdatedAt(),
-                article.getViews(),
-                DocumentTypeResponse.fromEntity(article.getDocumentType()),
-                SubjectDomainResponse.fromEntity(article.getSubjectDomain()),
-                article.getFiles().stream()
-                        .map(FileResponse::fromEntity)
-                        .collect(Collectors.toList())
-        );
-    }
-}
