--- src/main/java/re/kr/icuh/icuhplatformadmin/dto/DocumentTypeResponse.java
@@ -1,15 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.dto;
-
-import re.kr.icuh.icuhplatformadmin.domain.article.DocumentType;
-
-public record DocumentTypeResponse(
-        Long id,
-        String name
-) {
-    public static DocumentTypeResponse fromEntity(DocumentType documentType) {
-        return new DocumentTypeResponse(
-                documentType.getId(),
-                documentType.getName()
-        );
-    }
-}
