--- src/main/java/re/kr/icuh/icuhplatformadmin/dto/FileResponse.java
@@ -1,25 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.dto;
-
-import re.kr.icuh.icuhplatformadmin.domain.file.FileEntity;
-
-import java.time.LocalDateTime;
-
-public record FileResponse(
-        Long id,
-        String originalFilename,
-        String filePath,
-        Long fileSize,
-        String extension,
-        LocalDateTime createdAt
-) {
-    public static FileResponse fromEntity(FileEntity file) {
-        return new FileResponse(
-                file.getId(),
-                file.getOriginalFilename(),
-                file.getFilePath(),
-                file.getFileSize(),
-                file.getExtension(),
-                file.getCreatedAt()
-        );
-    }
-}
