--- src/main/java/re/kr/icuh/icuhplatformadmin/dto/SubjectDomainResponse.java
@@ -1,15 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.dto;
-
-import re.kr.icuh.icuhplatformadmin.domain.article.SubjectDomain;
-
-public record SubjectDomainResponse(
-        Long id,
-        String name
-) {
-    public static SubjectDomainResponse fromEntity(SubjectDomain subjectDomain) {
-        return new SubjectDomainResponse(
-                subjectDomain.getId(),
-                subjectDomain.getName()
-        );
-    }
-}
