+++ src/main/java/re/kr/icuh/icuhplatformadmin/file/domain/FileStatus.java
@@ -1,0 +1,10 @@
+package re.kr.icuh.icuhplatformadmin.file.domain;
+
+public enum FileStatus {
+    PENDING,
+    APPROVED,
+    REJECTED,
+    DELETED,
+    UPDATED_PENDING,
+    DELETED_PENDING
+}
