+++ src/main/java/re/kr/icuh/icuhplatformadmin/file/infrastructure/jpa/FileQueryRepository.java
@@ -1,0 +1,25 @@
+package re.kr.icuh.icuhplatformadmin.file.infrastructure.jpa;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.stereotype.Repository;
+import re.kr.icuh.icuhplatformadmin.file.domain.QFileEntity;
+
+@Repository
+public class FileQueryRepository {
+
+    private final JPAQueryFactory queryFactory;
+
+    public FileQueryRepository(JPAQueryFactory queryFactory) {
+        this.queryFactory = queryFactory;
+    }
+
+    public Long findFileByArticleId(Long articleId) {
+        QFileEntity file = QFileEntity.fileEntity;
+
+        return queryFactory
+                .select(file.id)
+                .from(file)
+                .where(file.article.id.eq(articleId))
+                .fetchOne();
+    }
+}
