+++ src/main/java/re/kr/icuh/icuhplatformadmin/service/ArticleService.java
@@ -3,13 +3,13 @@ package re.kr.icuh.icuhplatformadmin.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import re.kr.icuh.icuhplatformadmin.domain.article.Article;
-import re.kr.icuh.icuhplatformadmin.domain.article.ArticleEditRequest;
-import re.kr.icuh.icuhplatformadmin.domain.article.ArticleStatus;
-import re.kr.icuh.icuhplatformadmin.domain.file.FileStatus;
-import re.kr.icuh.icuhplatformadmin.dto.ArticleListResponse;
-import re.kr.icuh.icuhplatformadmin.dto.ArticlePendingResponse;
-import re.kr.icuh.icuhplatformadmin.repository.ArticleQueryRepository;
+import re.kr.icuh.icuhplatformadmin.article.domain.Article;
+import re.kr.icuh.icuhplatformadmin.article.domain.ArticleEditRequest;
+import re.kr.icuh.icuhplatformadmin.article.domain.ArticleStatus;
+import re.kr.icuh.icuhplatformadmin.file.domain.FileStatus;
+import re.kr.icuh.icuhplatformadmin.article.api.response.ArticleListResponse;
+import re.kr.icuh.icuhplatformadmin.article.api.response.ArticlePendingResponse;
+import re.kr.icuh.icuhplatformadmin.article.infrastructure.jpa.ArticleQueryRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
